Move common basic auth functionality to a separate module
