+++ basic-auth/pom.xml
@@ -9,34 +9,25 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>jwt</artifactId>
+    <artifactId>basic-auth</artifactId>
 
     <dependencies>
         <dependency>
             <groupId>com.authguard</groupId>
             <artifactId>service-api</artifactId>
             <version>${project.version}</version>
-            <scope>provided</scope>
         </dependency>
 
         <dependency>
             <groupId>com.authguard</groupId>
             <artifactId>config</artifactId>
             <version>${project.version}</version>
-            <scope>provided</scope>
         </dependency>
 
         <dependency>
             <groupId>com.authguard</groupId>
             <artifactId>dal</artifactId>
             <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>com.auth0</groupId>
-            <artifactId>java-jwt</artifactId>
-            <version>${auth0-jwt.version}</version>
         </dependency>
     </dependencies>
