+++ basic-auth/src/main/java/com/authguard/basic/passwords/AbstractSecurePassword.java
@@ -1,4 +1,4 @@
-package com.authguard.service.passwords;
+package com.authguard.basic.passwords;
 
 import com.authguard.service.model.HashedPasswordBO;
 import com.authguard.service.random.CryptographicRandom;
