+++ basic-auth/src/main/java/com/authguard/basic/passwords/BCryptPassword.java
@@ -1,4 +1,4 @@
-package com.authguard.service.passwords;
+package com.authguard.basic.passwords;
 
 import com.authguard.config.ConfigContext;
 import com.authguard.service.config.BCryptConfig;
