+++ basic-auth/src/main/java/com/authguard/basic/passwords/SecurePassword.java
@@ -1,4 +1,4 @@
-package com.authguard.service.passwords;
+package com.authguard.basic.passwords;
 
 import com.authguard.service.model.HashedPasswordBO;
