+++ basic-auth/src/main/java/com/authguard/basic/passwords/Violation.java
@@ -1,4 +1,4 @@
-package com.authguard.service.passwords;
+package com.authguard.basic.passwords;
 
 public class Violation {
     public enum Type {
