+++ basic-auth/src/test/java/com/authguard/basic/passwords/PasswordCharacteristicsTest.java
@@ -1,4 +1,4 @@
-package com.authguard.service.passwords;
+package com.authguard.basic.passwords;
 
 import org.junit.jupiter.api.Test;
