+++ basic-auth/src/test/java/com/authguard/basic/passwords/SCryptPasswordTest.java
@@ -1,18 +1,17 @@
-package com.authguard.service.passwords;
+package com.authguard.basic.passwords;
 
-import com.authguard.service.passwords.BCryptPassword;
 import org.apache.commons.lang3.RandomStringUtils;
 import com.authguard.service.model.HashedPasswordBO;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class BCryptPasswordTest {
+class SCryptPasswordTest {
     @Test
     void hash() {
-        final BCryptPassword bcrypt = new BCryptPassword();
+        final SCryptPassword scrypt = new SCryptPassword();
         final String password = RandomStringUtils.randomAlphanumeric(12);
-        final HashedPasswordBO hashedPassword = bcrypt.hash(password);
+        final HashedPasswordBO hashedPassword = scrypt.hash(password);
 
         assertThat(hashedPassword).isNotNull();
         assertThat(hashedPassword.getPassword()).isNotEqualTo(password);
@@ -21,7 +20,7 @@ class BCryptPasswordTest {
 
     @Test
     void verify() {
-        final BCryptPassword scrypt = new BCryptPassword();
+        final SCryptPassword scrypt = new SCryptPassword();
         final String password = RandomStringUtils.randomAlphanumeric(12);
         final HashedPasswordBO hashedPassword = scrypt.hash(password);
 
@@ -30,7 +29,7 @@ class BCryptPasswordTest {
 
     @Test
     void verifyMismatch() {
-        final BCryptPassword scrypt = new BCryptPassword();
+        final SCryptPassword scrypt = new SCryptPassword();
         final String password = RandomStringUtils.randomAlphanumeric(12);
         final HashedPasswordBO hashedPassword = scrypt.hash(password);
