+++ service-api/src/main/java/com/authguard/service/ApplicationsService.java
