+++ service-api/src/main/java/com/authguard/service/IdempotencyService.java
