+++ service/src/main/java/com/authguard/service/exchange/BasicToIdToken.java
@@ -1,6 +1,6 @@
 package com.authguard.service.exchange;
 
-import com.authguard.service.basic.BasicAuthProvider;
+import com.authguard.basic.BasicAuthProvider;
 import com.authguard.jwt.IdTokenProvider;
 import com.authguard.service.model.TokensBO;
 import com.google.inject.Inject;
