+++ service/src/main/java/com/authguard/service/exchange/BasicToPasswordless.java
@@ -1,6 +1,6 @@
 package com.authguard.service.exchange;
 
-import com.authguard.service.basic.BasicAuthProvider;
+import com.authguard.basic.BasicAuthProvider;
 import com.authguard.service.model.TokensBO;
 import com.authguard.service.passwordless.PasswordlessProvider;
 import com.google.inject.Inject;
