+++ service/src/test/java/com/authguard/service/impl/CredentialsServiceImplTest.java
@@ -10,14 +10,14 @@ import com.authguard.service.IdempotencyService;
 import com.authguard.service.config.PasswordConditions;
 import com.authguard.service.config.PasswordsConfig;
 import com.authguard.service.exceptions.ServiceException;
-import com.authguard.service.passwords.ServiceInvalidPasswordException;
+import com.authguard.basic.passwords.ServiceInvalidPasswordException;
 import com.authguard.service.mappers.ServiceMapperImpl;
 import com.authguard.service.model.AccountBO;
 import com.authguard.service.model.CredentialsBO;
 import com.authguard.service.model.HashedPasswordBO;
 import com.authguard.service.model.RequestContextBO;
-import com.authguard.service.passwords.PasswordValidator;
-import com.authguard.service.passwords.SecurePassword;
+import com.authguard.basic.passwords.PasswordValidator;
+import com.authguard.basic.passwords.SecurePassword;
 import org.jeasy.random.EasyRandom;
 import org.jeasy.random.EasyRandomParameters;
 import org.junit.jupiter.api.BeforeAll;
